Move registry access to BrowserPicker.Windows.dll, partially adresses issue #49, but FilePicker is still used in frontend.
