+++ BrowserPicker.sln
@@ -16,6 +16,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BrowserPicker", "src\Browse
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BrowserPicker.App", "src\BrowserPicker.App\BrowserPicker.App.csproj", "{B875AE86-5212-4F7F-BB1C-2BAA1FC110BC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BrowserPicker.Windows", "src\BrowserPicker.Windows\BrowserPicker.Windows.csproj", "{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -60,6 +62,18 @@ Global
 		{B875AE86-5212-4F7F-BB1C-2BAA1FC110BC}.SignedRelease|Any CPU.Build.0 = SignedRelease|Any CPU
 		{B875AE86-5212-4F7F-BB1C-2BAA1FC110BC}.SignedRelease|x86.ActiveCfg = SignedRelease|Any CPU
 		{B875AE86-5212-4F7F-BB1C-2BAA1FC110BC}.SignedRelease|x86.Build.0 = SignedRelease|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Debug|x86.Build.0 = Debug|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Release|x86.ActiveCfg = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.Release|x86.Build.0 = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.SignedRelease|Any CPU.ActiveCfg = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.SignedRelease|Any CPU.Build.0 = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.SignedRelease|x86.ActiveCfg = Release|Any CPU
+		{C3CE17EA-BAAE-4DE7-AFB2-319D50ECB2C8}.SignedRelease|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
