+++ src/BrowserPicker.App/ViewModel/ApplicationViewModel.cs
@@ -13,7 +13,6 @@ using System.Windows.Input;
 using BrowserPicker.Framework;
 using BrowserPicker.View;
 using JetBrains.Annotations;
-using Microsoft.Win32;
 
 namespace BrowserPicker.ViewModel
 {
@@ -25,7 +24,7 @@ namespace BrowserPicker.ViewModel
 		{
 			UnderlyingTargetURL = "https://github.com/mortenn/BrowserPicker";
 			force_choice = true;
-			Configuration = AppSettings.Settings;
+			Configuration = App.Settings;
 			Choices = new ObservableCollection<BrowserViewModel>(WellKnownBrowsers.List.Select(b => new BrowserViewModel(new BrowserModel(b, null, null), this)));
 		}
 
@@ -36,7 +35,7 @@ namespace BrowserPicker.ViewModel
 			UnderlyingTargetURL = TargetURL;
 			force_choice = options.Contains("/choose");
 			ConfigurationMode = TargetURL == null;
-			Configuration = AppSettings.Settings;
+			Configuration = App.Settings;
 			Choices = new ObservableCollection<BrowserViewModel>(Configuration.BrowserList.Select(m => new BrowserViewModel(m, this)));
 		}
 
@@ -44,7 +43,7 @@ namespace BrowserPicker.ViewModel
 		{
 			if (Choices.Count == 0 || (DateTime.Now - Configuration.LastBrowserScanTime) > TimeSpan.FromDays(7) || Choices.All(c => c.Model.PrivacyArgs == null))
 			{
-				FindBrowsers();
+				Configuration.FindBrowsers();
 			}
 			if (Configuration.AlwaysPrompt || ConfigurationMode || force_choice)
 			{
@@ -63,7 +62,7 @@ namespace BrowserPicker.ViewModel
 
 		public async Task ScanURLAsync(CancellationToken token)
 		{
-			var url = new UrlHandler(TargetURL);
+			var url = new UrlHandler(Configuration, TargetURL);
 			try
 			{
 				await url.ScanURLAsync(token);
@@ -100,7 +99,7 @@ namespace BrowserPicker.ViewModel
 			start.Select.Execute(null);
 		}
 
-		public ICommand RefreshBrowsers => new DelegateCommand(FindBrowsers);
+		public ICommand RefreshBrowsers => new DelegateCommand(Configuration.FindBrowsers);
 
 		public ICommand Configure => new DelegateCommand(() => ConfigurationMode = !ConfigurationMode);
 
@@ -133,7 +132,7 @@ namespace BrowserPicker.ViewModel
 			Configuration.AddBrowser(browser.Model);
 		}
 
-		public AppSettings Configuration { get; }
+		public IBrowserPickerConfiguration Configuration { get; }
 
 		public ObservableCollection<BrowserViewModel> Choices { get; }
 
@@ -233,97 +232,6 @@ namespace BrowserPicker.ViewModel
 		{
 			EditURL = UnderlyingTargetURL;
 			OnPropertyChanged(nameof(EditURL));
-		}
-
-		private void FindBrowsers()
-		{
-			var removed = Choices.Where(b => b.Model.Removed).ToList();
-			if (removed.Count > 0)
-				removed.ForEach(b => Choices.Remove(b));
-
-			// Prefer 64 bit browsers to 32 bit ones, machine wide installs to user specific ones.
-			EnumerateBrowsers(Registry.LocalMachine, @"SOFTWARE\Clients\StartMenuInternet");
-			EnumerateBrowsers(Registry.CurrentUser, @"SOFTWARE\Clients\StartMenuInternet");
-			EnumerateBrowsers(Registry.LocalMachine, @"SOFTWARE\WOW6432Node\Clients\StartMenuInternet");
-			EnumerateBrowsers(Registry.CurrentUser, @"SOFTWARE\WOW6432Node\Clients\StartMenuInternet");
-
-			if (!Choices.Any(browser => browser.Model.Name.Contains("Edge")))
-			{
-				FindLegacyEdge();
-			}
-
-			Configuration.LastBrowserScanTime = DateTime.Now;
-		}
-
-		/// <summary>
-		/// This is used to detect the old Edge browser.
-		/// If the computer has the new Microsoft Edge browser installed, this should never be called.
-		/// </summary>
-		private void FindLegacyEdge()
-		{
-			var systemApps = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Windows), "SystemApps");
-			if (!Directory.Exists(systemApps))
-				return;
-
-			var targets = Directory.GetDirectories(systemApps, "*MicrosoftEdge_*");
-			if (targets.Length <= 0)
-			{
-				return;
-			}
-			var known = WellKnownBrowsers.Lookup("Edge", null);
-			var appId = Path.GetFileName(targets[0]);
-			var icon = Path.Combine(targets[0], "Assets", "MicrosoftEdgeSquare44x44.targetsize-32_altform-unplated.png");
-			var shell = $"shell:AppsFolder\\{appId}!MicrosoftEdge";
-
-			var model = new BrowserModel(known, icon, shell);
-			AddOrUpdateBrowserModel(model);
-		}
-
-		private void EnumerateBrowsers(RegistryKey hive, string subKey)
-		{
-			var root = hive.OpenSubKey(subKey, false);
-			if (root == null)
-				return;
-			foreach (var browser in root.GetSubKeyNames().Where(n => n != "BrowserPicker"))
-				GetBrowserDetails(root, browser);
-		}
-
-		private void GetBrowserDetails(RegistryKey root, string browser)
-		{
-			var reg = root.OpenSubKey(browser, false);
-			if (reg == null)
-				return;
-
-			var (name, icon, shell) = reg.GetBrowser();
-
-			if (string.IsNullOrWhiteSpace(name) || string.IsNullOrWhiteSpace(shell))
-			{
-				return;
-			}
-			var known = WellKnownBrowsers.Lookup(name, shell);
-			if (known != null)
-			{
-				var knownModel = new BrowserModel(known, icon, shell);
-				AddOrUpdateBrowserModel(knownModel);
-				return;
-			}
-			var model = new BrowserModel(name, icon, shell);
-			AddOrUpdateBrowserModel(model);
-		}
-
-		private void AddOrUpdateBrowserModel(BrowserModel model)
-		{
-			var update = Configuration.BrowserList.FirstOrDefault(m => m.Name.Equals(model.Name, StringComparison.CurrentCultureIgnoreCase));
-			if (update != null)
-			{
-				update.Command = model.Command;
-				update.CommandArgs = model.CommandArgs;
-				update.PrivacyArgs = model.PrivacyArgs;
-				update.IconPath = model.IconPath;
-				return;
-			}
-			Choices.Add(new BrowserViewModel(model, this));
-			Configuration.AddBrowser(model);
 		}
 
 		public event PropertyChangedEventHandler PropertyChanged;
