+++ src/BrowserPicker.Windows/RegistryHelpers.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
 
-namespace BrowserPicker.Framework
+namespace BrowserPicker.Windows
 {
 	public static class RegistryHelpers
 	{
